+++ ios/BDD_IOS/.gitignore
@@ -1,0 +1,5 @@
+allure-results/
+screenshots/
+logs/
+data/
+pages/__pycache__/
