+++ ios/BDD_IOS/README.md
@@ -1,0 +1,44 @@
+## **Python with Selenium Behave BDD Framework**
+
+This is a BDD automation framework developed on Selenium and Python Behave.
+Behave BDD framework details can be found: https://behave.readthedocs.io/en/stable/ and basic tutorial: https://behave.readthedocs.io/en/stable/tutorial.html
+
+To reduce the duplication, Page Object Model is used in this framework.
+Page Object Model, also known as POM, is a design pattern in Selenium that creates an object repository for storing all web elements. It is useful in reducing code duplication and improves test case maintenance.
+In Page Object Model, consider each web page of an application as a class file. Each class file will contain only corresponding web page elements. Using these elements, testers can perform operations on the website under test.
+
+**Pages** folder contains the elements and corresponding actions of the pages. BasePage.py contains all functions those are required to be call multiple times from different project to reduce the duplicate code.
+
+**features** folder contains **steps** folder which has all the test files and also the feature files.
+
+**reports** directory contains the json files generated by allure reports
+
+**Utilities** directory contains the customLogger.py and readConfig.py to help get value from config.ini file along with log generation.
+
+**Logs** custom logger implemented for logs of test execution, all logs file can be found on Logs folder.
+
+**requirements.txt** file contains all the python packages needed to run this framework
+
+**PyCharm IDE (Community Edition) is used.**
+
+**Python version - Python 3.10.5**
+
+### **Commands to run the tests**
+
+**To run the test with allure report**
+`behave -f allure_behave.formatter:AllureFormatter -o reports/ features/XXYYZZ.feature`
+
+**To run the test without allure report** 
+`behave features/XXYYZZ.feature`
+
+**To run the particular scenario**
+`behave -n "scenario name"`
+
+
+**To generate the html allure report from the json files inside Reports folder**
+`allure serve Reports/`
+
+
+**To generate the html allure report from the json files inside Reports folder and open the report in browser**
+`/opt/allure-2.24.1/bin/allure generate reports/allure_result -o reports/allure_report --clean`
+
