+++ ios/BDD_IOS/behave.ini
@@ -1,0 +1,5 @@
+[behave]
+color = yes
+stdout_capture = yes
+stderr_capture = yes
+log_capture = no
