+++ ios/BDD_IOS/configuration/config.ini
@@ -2,6 +2,10 @@
 # environment options : [test]
 environment = test
 
+# Hardcoded configuration values
+package_name = "com.instagram.android"
+activity = "com.instagram.android.activity.MainTabActivity"
+
 # browser options : [edge, chrome, firefox]
 browser_mobile = Safari
 
@@ -12,5 +16,4 @@ headless = False
 loglevel = ERROR
 
 [test]
-ssp_url = https://ss-uhcwuat.staffpassports.nhs.uk/
 hr_url = https://hr-uhcwuat.staffpassports.nhs.uk/
