+++ ios/BDD_IOS/features/steps/dsp_hr_portal_archive_credential_steps.py
@@ -4,8 +4,9 @@ Steps to revoke credentials in the wallet and validate its success by the HR use
 
 from behave import given, when, then
 
-from ios.BDD_IOS.pages.hr_archive_credential import HRArchiveCredentialPage
-from ios.BDD_IOS.pages.base_page import BasePage
+from pages.hr_pages.hr_archive_credential import HRArchiveCredentialPage
+from pages.base_page import BasePage
+
 
 
 @then("HR User clicked on Fire safety credential which needs to archive")
