+++ ios/BDD_IOS/features/steps/dsp_hr_portal_identity_access_steps.py
@@ -3,8 +3,10 @@ Steps definition w.r.t (Identity & Access) features
 """
 from behave import given, when, then
 
-from ios.BDD_IOS.pages.hr_identity_and_access import HRIdentityAndAccessPage
-from ios.BDD_IOS.pages.base_page import BasePage
+from pages.hr_pages.hr_identity_and_access import HRIdentityAndAccessPage
+from pages.base_page import BasePage
+
+
 
 @then('Click on the Identity & Access Tab with HR Portal')
 def hr_portal_identity_access(context):
