+++ ios/BDD_IOS/features/steps/dsp_hr_portal_reviewed_staff_passport_steps.py
@@ -2,8 +2,8 @@
 Steps definition for Reviewed staff passport features
 """
 from behave import given, when, then
-from ios.BDD_IOS.pages.base_page import BasePage
-from ios.BDD_IOS.pages.hr_reviewed_staff_passport import HRReviewedStaffPassportPage
+from pages.base_page import BasePage
+from pages.hr_pages.hr_reviewed_staff_passport import HRReviewedStaffPassportPage
 
 
 @then('Click on the Reviewed Staff Passport Tab under HR Portal')
