+++ ios/BDD_IOS/features/steps/dsp_hr_portal_temporary_placement_credential_steps.py
@@ -2,8 +2,9 @@
 Steps definition w.r.t Temporary credentials provided
 """
 from behave import given, when, then
-from ios.BDD_IOS.pages.hr_temporary_placement_credential import HRTemporaryPlacementCredentialPage
-from ios.BDD_IOS.pages.base_page import BasePage
+from pages.hr_pages.hr_temporary_placement_credential import HRTemporaryPlacementCredentialPage
+from pages.base_page import BasePage
+
 
 
 @then('HR User clicked on Create temporary placement credential link')
