+++ ios/BDD_IOS/flake8.txt
@@ -1,0 +1,4 @@
+[flake8]
+ignore = E501,C901,W504
+max-complexity = 10
+per-file-ignores = __init__.py:F401
