--- ios/BDD_IOS/gitignore
@@ -1,1 +1,0 @@
-pages/__pycache__/
