--- ios/BDD_IOS/logs/aos_app.log
@@ -1,1 +1,0 @@
-
